[GSW-2004] Moved the Pages folder above the `/src` folder
