+++ packages/web/pages/_app.tsx
@@ -24,7 +24,7 @@ import GnoswapServiceProvider from "@providers/gnoswap-service-provider/GnoswapS
 import GnoswapThemeProvider from "@providers/gnoswap-theme-provider/GnoswapThemeProvider";
 import SnackbarProvider from "@providers/snackbar-provider/SnackbarProvider";
 
-import nextI18NextConfig from "../../next-i18next.config.js";
+import nextI18NextConfig from "../next-i18next.config";
 import Custom500 from "./500";
 
 function App({ Component, pageProps }: AppProps) {
