+++ packages/web/pages/earn/pool/remove/index.tsx
