+++ packages/web/pages/earn/pool/stake/index.tsx
