+++ packages/web/pages/privacy/index.tsx
