+++ packages/web/pages/terms/index.tsx
