+++ packages/web/tsconfig.json
@@ -14,7 +14,7 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
-    "baseUrl": ".",
+    "baseUrl": "./",
     "paths": {
       "@constants/*": ["src/constants/*"],
       "@containers/*": ["src/containers/*"],
